005993_kpi
the subtask doesn't get involve in kpi anymore, the parent task get the reporting for them. new boolean field to enable or disable the budget calculation on the task. Separation of invoiced_mode T&M and other, if not T&M: hiding  PC Review Budget and Carry Forward Budget , and changing calculation for Realized Budget and Invoiced Budget. In the timesheet tab "progress" is renamed "Consumed hours" to be more explicit. in the project view Budget Consumed is based on realized budget  (not on timesheet anymore